Refactored com.helliongames -> com.nekomaster1000 Simplified config Fixed build.gradle Updated Gradle wrapper version Updated ForgeGradle version
