+++ .gitignore
@@ -20,6 +20,7 @@ build
 # other
 eclipse
 run
+.vscode
 
 # Files from Forge MDK
 forge*changelog.txt
