+++ src/main/java/com/nekomaster1000/secondchanceforge/SecondChanceForge.java
@@ -1,15 +1,16 @@
-package com.helliongames.secondchanceforge;
+package com.nekomaster1000.secondchanceforge;
 
-import com.helliongames.secondchanceforge.config.ConfigHelper;
-import com.helliongames.secondchanceforge.config.ConfigHolder;
+import com.nekomaster1000.secondchanceforge.config.SecondChanceConfig;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
 
 @Mod(SecondChanceForge.MOD_ID)
 public class SecondChanceForge {
@@ -26,9 +27,6 @@ public class SecondChanceForge {
         SecondChanceSoundEvents.register(modEventBus);
 
         //Registering Configs
-        modLoadingContext.registerConfig(ModConfig.Type.COMMON, ConfigHolder.COMMON_SPEC);
-
-        //Baking Configs
-        ConfigHelper.bakeCommon(null);
+        modLoadingContext.registerConfig(ModConfig.Type.COMMON, SecondChanceConfig.CONFIG_SPEC);
     }
 }
