+++ src/main/java/com/nekomaster1000/secondchanceforge/SecondChanceSoundEvents.java
@@ -1,4 +1,4 @@
-package com.helliongames.secondchanceforge;
+package com.nekomaster1000.secondchanceforge;
 
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.SoundEvent;
