+++ src/main/java/com/nekomaster1000/secondchanceforge/mixin/MixinConnector.java
@@ -1,4 +1,4 @@
-package com.helliongames.secondchanceforge.mixin;
+package com.nekomaster1000.secondchanceforge.mixin;
 
 import org.spongepowered.asm.mixin.Mixins;
 import org.spongepowered.asm.mixin.connect.IMixinConnector;
@@ -6,6 +6,6 @@ import org.spongepowered.asm.mixin.connect.IMixinConnector;
 public class MixinConnector implements IMixinConnector {
     @Override
     public void connect() {
-        Mixins.addConfiguration("second-chance-forge.mixins.json");
+        Mixins.addConfiguration("secondchanceforge.mixins.json");
     }
 }
