+++ src/main/resources/META-INF/MANIFEST.mf
@@ -1,1 +1,1 @@
-MixinConnector: com.helliongames.secondchanceforge.mixin.MixinConnector
+MixinConnector: com.nekomaster1000.secondchanceforge.mixin.MixinConnector
