+++ src/main/resources/secondchanceforge.mixins.json
@@ -1,8 +1,8 @@
 {
   "required": true,
   "minVersion": "0.8",
-  "package": "com.helliongames.secondchanceforge.mixin",
-  "refmap": "second-chance-forge.refmap.json",
+  "package": "com.nekomaster1000.secondchanceforge.mixin",
+  "refmap": "secondchanceforge.refmap.json",
   "compatibilityLevel": "JAVA_8",
   "injectors": {
     "defaultRequire": 1
